SMTProgram tests

+++ src/main/kotlin/tools/aqua/konstraints/solvers/Z3/Z3ExpressionGenerator.kt
@@ -351,7 +351,7 @@ fun Expression<BVSort>.z3ify(context: Z3Context): Expr<BitVecSort> =
     }
 
 fun BVLiteral.z3ify(context: Z3Context): Expr<BitVecSort> =
-    context.context.mkBV(this.value, this.bits)
+    context.context.mkBV(this.value.toInt(), this.bits)
 
 fun BVConcat.z3ify(context: Z3Context): Expr<BitVecSort> =
     context.context.mkConcat(this.lhs.z3ify(context), this.rhs.z3ify(context))
